backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
PyMuPDF
opencv-python
numpy
Pillow
